backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blinderfit-backend"
version = "1.0.0"
description = "AI-powered personalized health coach backend"
requires-python = ">=3.10"
authors = [
    {name = "Blinderfit Team", email = "team@blinderfit.com"},
]
maintainers = [
    {name = "Blinderfit Team", email = "team@blinderfit.com"},
]
keywords = ["health", "fitness", "ai", "fastapi", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core FastAPI dependencies
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",

    # Firebase dependencies
    "firebase-admin==6.8.0",

    # Google AI (Gemini)
    "google-generativeai==0.8.3",

    # HTTP client
    "httpx==0.28.1",

    # Data processing
    "pandas==2.2.3",
    "numpy==2.1.3",

    # Authentication
    "pyjwt==2.10.1",
    "cryptography==43.0.3",

    # Environment and configuration
    "python-dotenv==1.0.1",

    # Validation
    "email-validator==2.2.0",

    # Image processing for multimodal
    "Pillow==11.0.0",

    # Speech processing for voice
    "speechrecognition==3.11.0",
    "pydub==0.25.1",

    # Web search integration
    "google-search-results==2.4.2",

    # Wearables integration
    "google-auth==2.36.0",
    "google-auth-oauthlib==1.2.1",

    # Caching
    "redis==5.2.0",

    # Async support
    "aiofiles==24.1.0",

    # Health and monitoring
    "prometheus-client==0.21.1",

    # Security
    "python-multipart==0.0.19",
    "bcrypt==4.2.1",

    # Rate limiting
    "slowapi==0.1.9",

    # Additional utilities for integrations
    "requests==2.32.3",
    "urllib3==2.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "pytest-cov==5.0.0",
    "httpx==0.28.1",  # For testing FastAPI
]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.42",
]
test = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-cov==5.0.0",
    "httpx==0.28.1",
]

[project.urls]
Homepage = "https://blinderfit.com"
Documentation = "https://docs.blinderfit.com"
Repository = "https://github.com/blinderchief/blinderfit"
Issues = "https://github.com/blinderchief/blinderfit/issues"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "pytest-cov==5.0.0",
    "httpx==0.28.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "--tb=short",
    "-v",
    "--asyncio-mode=auto"
]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]
known_third_party = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "firebase_admin",
    "google",
    "httpx",
    "pandas",
    "numpy",
    "pyjwt",
    "cryptography",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".venv",
    ".tox",
    "migrations",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "firebase_admin.*",
    "google.*",
    "speechrecognition.*",
    "pydub.*",
]
ignore_missing_imports = true